@@ -2,103 +2,110 @@
 // Copyright (c) PlaceholderCompany. All rights reserved.
 // </copyright>
 
-namespace RSBingoBot.Component_interaction_handlers
+namespace RSBingoBot.Component_interaction_handlers;
+
+using DSharpPlus;
+using DSharpPlus.Entities;
+using DSharpPlus.EventArgs;
+using RSBingoBot.Discord_event_handlers;
+using static RSBingoBot.InteractionMessageUtilities;
+
+/// <summary>
+/// Handles the interaction with the "Create team" button in the team-registration channel.
+/// </summary>
+internal class CreateTeamButtonHandler : ComponentInteractionHandler
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-    using DSharpPlus;
-    using DSharpPlus.Entities;
-    using DSharpPlus.EventArgs;
-    using RSBingo_Framework.Models;
-    using RSBingoBot.Discord_event_handlers;
+    private const string InvalidNameCharactersMessage = "A team name must only contain letters and/or numbers.";
+    private const string NameTooLongMessage = "A team name cannot exceed {0} characters.";
+
+    private readonly string modalId = Guid.NewGuid().ToString();
+    private readonly string teamNameInputId = Guid.NewGuid().ToString();
 
     /// <summary>
-    /// Handles the interaction with the "Team sign-up" select menu in create-team channel.
+    /// Gets the custom Id for the "Create team" button.
     /// </summary>
-    internal class CreateTeamButtonHandler : ComponentInteractionHandler
+    public static string CreateTeamButtonId { get; } = "create-team-button";
+
+    /// <inheritdoc/>
+    protected override bool ContinueWithNullUser => true;
+    protected override bool CreateAutoResponse => false;
+    protected override bool AutoRegisterInteraction => false;
+
+    /// <inheritdoc/>
+    public async override Task InitialiseAsync(ComponentInteractionCreateEventArgs args, InitialisationInfo info)
     {
-        private readonly string modalId = Guid.NewGuid().ToString();
-        private readonly string teamNameInputId = Guid.NewGuid().ToString();
+        await base.InitialiseAsync(args, info);
 
-        /// <summary>
-        /// Gets the custom Id for the "Create team" button.
-        /// </summary>
-        public static string CreateTeamButtonId { get; } = "create-team-button";
-
-        /// <inheritdoc/>
-        protected override bool ContinueWithNullUser { get { return true; } }
-
-        protected override bool CreateAutoResponse { get { return false; } }
-
-        /// <inheritdoc/>
-        public async override Task InitialiseAsync(ComponentInteractionCreateEventArgs args, InitialisationInfo info)
+        if (DataWorker.Users.Exists(args.Interaction.User.Id))
         {
-            await base.InitialiseAsync(args, info);
-
-            var teamNameInput = new TextInputComponent("Team name", teamNameInputId);
-
-            var builder = new DiscordInteractionResponseBuilder()
-                .WithTitle("Create a team.")
-                .WithCustomId(modalId)
-                .AddComponents(teamNameInput);
-
-            await args.Interaction.CreateResponseAsync(InteractionResponseType.Modal, builder);
-            SubscribeModal(new ModalSubmittedDEH.Constraints(customId: modalId), TeamNameSubmitted);
-        }
-
-        private async Task TeamNameSubmitted(DiscordClient discordClient, ModalSubmitEventArgs args)
-        {
-            User = DataWorker.Users.GetByDiscordId(args.Interaction.User.Id);
-            await args.Interaction.CreateResponseAsync(InteractionResponseType.ChannelMessageWithSource, new DiscordInteractionResponseBuilder().WithContent("23"));
-            if (await TrySendUserTeamStatusErrorMessage(args.Interaction.User.Id, false, args) is false)
-            {
-                await ConcludeInteraction();
-            }
-
-            string teamName = args.Values[teamNameInputId];
-            var builder = new DiscordFollowupMessageBuilder()
-                .AsEphemeral();
-
-            if (DataWorker.Teams.DoesTeamExist(teamName))
-            {
-                builder.WithContent("A team with this name already exists.");
-                await args.Interaction.CreateFollowupMessageAsync(builder);
-            }
-            else
-            {
-                builder.WithContent("Creating team...");
-                DiscordMessage? followupMessage = await args.Interaction.CreateFollowupMessageAsync(builder);
-
-                RSBingoBot.DiscordTeam team = new(discordClient, teamName);
-                await team.InitialiseAsync(null);
-
-                User = DataWorker.Users.Create(args.Interaction.User.Id, teamName);
-                DataWorker.SaveChanges();
-
-                DiscordRole? role = await args.Interaction.Guild.CreateRoleAsync(teamName);
-                await args.Interaction.Guild.GetMemberAsync(args.Interaction.User.Id).Result.GrantRoleAsync(role);
-
-                string content = $"The team '{teamName}' has been created successfully.";
-
-                if (args.Interaction.GetFollowupMessageAsync(followupMessage.Id).Result != null)
-                {
-                    var editBuilder = new DiscordWebhookBuilder()
-                        .WithContent(content);
-                    await args.Interaction.EditFollowupMessageAsync(followupMessage.Id, editBuilder);
-                }
-                else
-                {
-                    // TODO: Figure out why this doesn't do anything.
-                    var newBuilder = new DiscordFollowupMessageBuilder()
-                        .WithContent(content)
-                        .AsEphemeral();
-                    await args.Interaction.CreateFollowupMessageAsync(newBuilder);
-                }
-            }
+            await Respond(args, AlreadyOnATeamMessage, true);
             await ConcludeInteraction();
+            return;
         }
+
+        await ModalResponse();
+    }
+
+    private async Task ModalResponse()
+    {
+        var teamNameInput = new TextInputComponent("Team name", teamNameInputId);
+
+        var builder = new DiscordInteractionResponseBuilder()
+            .WithTitle("Create a team.")
+            .WithCustomId(modalId)
+            .AddComponents(teamNameInput);
+
+        await CurrentInteractionArgs.Interaction.CreateResponseAsync(InteractionResponseType.Modal, builder);
+        SubscribeModal(new ModalSubmittedDEH.Constraints(customId: modalId), TeamNameSubmitted);
+    }
+
+    private async Task TeamNameSubmitted(DiscordClient discordClient, ModalSubmitEventArgs args)
+    {
+        RegisterUserInstance();
+        await ProcessRequest(discordClient, args);
+        await ConcludeInteraction();
+    }
+
+    private async Task ProcessRequest(DiscordClient discordClient, ModalSubmitEventArgs args)
+    {
+        string teamName = args.Values[teamNameInputId];
+        string errorMessage = GetNameErrorMessage(teamName);
+
+        if (string.IsNullOrEmpty(errorMessage) is false)
+        {
+            await Respond(args, errorMessage, true);
+            return;
+        }
+
+        if (DataWorker.Teams.DoesTeamExist(teamName))
+        {
+            await Respond(args, "A team with this name already exists.", true);
+            return;
+        }
+
+        await CreateTeam(discordClient, args, teamName);
+    }
+
+    private async Task CreateTeam(DiscordClient discordClient, ModalSubmitEventArgs args, string teamName)
+    {
+        await Respond(args, "Creating team...", true);
+
+        RSBingoBot.DiscordTeam team = new(discordClient, teamName);
+        await team.InitialiseAsync();
+
+        User = DataWorker.Users.Create(args.Interaction.User.Id, teamName);
+        DataWorker.SaveChanges();
+
+        await args.Interaction.Guild.GetMemberAsync(args.Interaction.User.Id).Result.GrantRoleAsync(team.Role);
+
+        await EditResponse(args, $"The team '{teamName}' has been created successfully.", true);
+    }
+
+    private string GetNameErrorMessage(string name)
+    {
+        string errorMessage = "";
+        if (name.Any(ch => char.IsLetterOrDigit(ch) is false)) { errorMessage += InvalidNameCharactersMessage + Environment.NewLine; }
+        if (name.Length > General.TeamNameMaxLength) { errorMessage += NameTooLongMessage.FormatConst(General.TeamNameMaxLength); }
+        return errorMessage;
     }
 }
