@@ -2,129 +2,127 @@
 // Copyright (c) PlaceholderCompany. All rights reserved.
 // </copyright>
 
-namespace RSBingoBot.Component_interaction_handlers
+namespace RSBingoBot.Component_interaction_handlers;
+
+using RSBingoBot.Discord_event_handlers;
+using RSBingo_Framework.Models;
+using DSharpPlus;
+using DSharpPlus.Entities;
+using DSharpPlus.EventArgs;
+using static RSBingoBot.InteractionMessageUtilities;
+
+/// <summary>
+/// Handles the interaction with the "Join team" button.
+/// </summary>
+internal class JoinTeamButtonHandler : ComponentInteractionHandler
 {
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Threading.Tasks;
-    using DSharpPlus;
-    using DSharpPlus.Entities;
-    using DSharpPlus.EventArgs;
-    using RSBingo_Framework.Models;
-    using RSBingoBot.Discord_event_handlers;
+    private readonly string confirmButtonId = Guid.NewGuid().ToString();
+    private readonly string teamSelectId = Guid.NewGuid().ToString();
+    private string teamSelected = string.Empty;
+
+    /// <inheritdoc/>
+    protected override bool ContinueWithNullUser { get { return true; } }
+    protected override bool CreateAutoResponse { get { return false; } }
 
     /// <summary>
-    /// Handles the interaction with the "Team sign-up" select menu in create-team channel.
+    /// Gets the custom Id for the "Join team" button.
     /// </summary>
-    internal class JoinTeamButtonHandler : ComponentInteractionHandler
+    public static string JoinTeamButtonId { get; } = "join-team-button";
+
+    /// <inheritdoc/>
+    public async override Task InitialiseAsync(ComponentInteractionCreateEventArgs args, InitialisationInfo info)
     {
-        private readonly string confirmButtonId = Guid.NewGuid().ToString();
-        private readonly string teamSelectId = Guid.NewGuid().ToString();
-        private string teamSelected = string.Empty;
+        await base.InitialiseAsync(args, info);
 
-        /// <inheritdoc/>
-        protected override bool ContinueWithNullUser { get { return true; } }
-        protected override bool CreateAutoResponse { get { return true; } }
+        if (await ValidateUser()) { await TeamSelectionResponse(); }
+    }
 
-
-        /// <summary>
-        /// Gets the custom Id for the "Join team" button.
-        /// </summary>
-        public static string JoinTeamButtonId { get; } = "join-team-button";
-
-        /// <inheritdoc/>
-        public async override Task InitialiseAsync(ComponentInteractionCreateEventArgs args, InitialisationInfo info)
+    private async Task<bool> ValidateUser()
+    {
+        if (DataWorker.Users.Exists(OriginalInteractionArgs.Interaction.User.Id))
         {
-            await base.InitialiseAsync(args, info);
+            await Respond(OriginalInteractionArgs, AlreadyOnATeamMessage, true);
+            await ConcludeInteraction();
+            return false;
+        }
+        return true;
+    }
 
-            var confirmButton = new DiscordButtonComponent(ButtonStyle.Primary, confirmButtonId, "Confirm");
-            SubscribeComponent(
-                new ComponentInteractionDEH.Constraints(user: args.User, channel: args.Channel, customId: confirmButtonId),
-                TeamJoinConfirmed, true);
-
-            var builder = new DiscordFollowupMessageBuilder();
-            IEnumerable<Team> teams = DataWorker.Teams.GetTeams();
-
-            if (await TrySendUserTeamStatusErrorMessage(args.User.Id, false, args) is false)
-            {
-                await ConcludeInteraction();
-            }
-
-            if (!teams.Any())
-            {
-                builder
-                   .WithContent("No teams created.")
-                   .AsEphemeral();
-            }
-            else
-            {
-                var options = new List<DiscordSelectComponentOption>();
-                foreach (Team team in teams)
-                {
-                    options.Add(new (team.Name, team.Name));
-                }
-
-                var teamSelect = new DiscordSelectComponent(teamSelectId, "Select team", options);
-                SubscribeComponent(
-                    new ComponentInteractionDEH.Constraints(user: args.User, channel: args.Channel, customId: teamSelectId),
-                    TeamSelected, true);
-
-                builder
-                    .WithContent($"{args.User.Mention} Select a team to join.")
-                    .AddComponents(teamSelect)
-                    .AddComponents(confirmButton);
-            }
-            DiscordMessage? followupMessage = await args.Interaction.CreateFollowupMessageAsync(builder);
-            MessagesForCleanup.Add(followupMessage);
+    private async Task TeamSelectionResponse()
+    {
+        if (await TrySendUserTeamStatusErrorMessage(OriginalInteractionArgs.User.Id, false, OriginalInteractionArgs))
+        {
+            await ConcludeInteraction();
+            return;
         }
 
-        private async Task TeamSelected(DiscordClient discordClient, ComponentInteractionCreateEventArgs args)
+        IEnumerable<Team> teams = DataWorker.Teams.GetTeams();
+
+        if (teams.Any() is false)
         {
-            teamSelected = args.Values[0];
-            await args.Interaction.CreateResponseAsync(InteractionResponseType.DeferredMessageUpdate);
+            await Respond(OriginalInteractionArgs, "No teams created.", true);
+            await ConcludeInteraction();
+            return;
         }
 
-        private async Task TeamJoinConfirmed(DiscordClient discordClient, ComponentInteractionCreateEventArgs args)
+        IEnumerable<DiscordSelectComponentOption> options = teams.Select(t => new DiscordSelectComponentOption(t.Name, t.Name));
+
+        DiscordSelectComponent teamSelect = new(teamSelectId, "Select team", options);
+        SubscribeComponent(
+            new ComponentInteractionDEH.Constraints(user: OriginalInteractionArgs.User, channel: OriginalInteractionArgs.Channel, customId: teamSelectId),
+            TeamSelected, true);
+
+        DiscordButtonComponent confirmButton = new(ButtonStyle.Primary, confirmButtonId, "Confirm");
+        SubscribeComponent(
+            new ComponentInteractionDEH.Constraints(user: OriginalInteractionArgs.User, channel: OriginalInteractionArgs.Channel, customId: confirmButtonId),
+            TeamJoinConfirmed, true);
+
+        var builder = new DiscordInteractionResponseBuilder()
+            .WithContent($"{OriginalInteractionArgs.User.Mention} Select a team to join.")
+            .AddComponents(teamSelect)
+            .AddComponents(confirmButton);
+
+        await OriginalInteractionArgs.Interaction.CreateResponseAsync(InteractionResponseType.ChannelMessageWithSource, builder);
+        MessagesForCleanup.Add(await OriginalInteractionArgs.Interaction.GetOriginalResponseAsync());
+    }
+
+    private async Task TeamSelected(DiscordClient discordClient, ComponentInteractionCreateEventArgs args) =>
+        teamSelected = args.Values[0];
+
+    private async Task TeamJoinConfirmed(DiscordClient discordClient, ComponentInteractionCreateEventArgs args)
+    {
+        if (teamSelected == string.Empty)
         {
-            string content = string.Empty;
-
-            if (await TrySendUserTeamStatusErrorMessage(args.User.Id, false, args) is false)
-            {
-                await ConcludeInteraction();
-            }
-
-            if (teamSelected == string.Empty)
-            {
-                content = "You must select a team to join.";
-            }
-            else
-            {
-                DataWorker.Users.Create(args.User.Id, teamSelected);
-                DataWorker.SaveChanges();
-
-                var roles = args.Guild.Roles.Select(x => x.Value).ToList();
-                int index = roles.Select(x => x.Name).ToList().IndexOf(teamSelected);
-
-                content = $"You have joined team {teamSelected}.";
-
-                if (index == -1)
-                {
-                    // Error, team role should exist
-                    content += $"{Environment.NewLine}The team's role does not exist; please tell an admin.";
-                }
-                else
-                {
-                    await args.Guild.GetMemberAsync(args.User.Id).Result.GrantRoleAsync(roles[index]);
-                }
-
-                await ConcludeInteraction();
-            }
-
-            var builder = new DiscordFollowupMessageBuilder()
-                .WithContent(content)
-                .AsEphemeral();
-
-            await args.Interaction.CreateFollowupMessageAsync(builder);
+            await Followup(args, "You must select a team to join.", true);
+            return;
         }
+
+        await JoinTeam(args);
+    }
+
+    private async Task JoinTeam(ComponentInteractionCreateEventArgs args)
+    {
+        string content = string.Empty;
+
+        Team? team = DataWorker.Teams.GetByName(teamSelected);
+
+        DataWorker.Users.Create(args.User.Id, team);
+        DataWorker.SaveChanges();
+
+        content = $"You have joined team {teamSelected}.";
+
+        DiscordRole role = RSBingoBot.DiscordTeam.GetInstance(team).Role;
+
+        try
+        {
+            await args.Guild.GetMemberAsync(args.User.Id).Result.GrantRoleAsync(role);
+        }
+        catch (DSharpPlus.Exceptions.NotFoundException)
+        {
+            content += $"{Environment.NewLine}The team's role does not exist; please tell a host.";
+        }
+
+        await Followup(args, content, true);
+        await ConcludeInteraction();
     }
 }
