@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using DSharpPlus;
 using DSharpPlus.Entities;
 using DSharpPlus.EventArgs;
-using Emzi0767.Utilities;
+//using Emzi0767.Utilities;
 using RSBingoBot.Interfaces;
 
 /// <summary>
