@@ -4,48 +4,63 @@
 
 namespace RSBingoBot;
 
-using DSharpPlus;
-using DSharpPlus.Entities;
 using RSBingoBot.Imaging;
 using RSBingoBot.Component_interaction_handlers;
 using RSBingoBot.Component_interaction_handlers.Testing;
 using RSBingo_Framework.Models;
 using RSBingo_Framework.Records;
 using RSBingo_Framework.Interfaces;
+using DSharpPlus;
+using DSharpPlus.Entities;
 using SixLabors.ImageSharp;
 using static RSBingo_Framework.DAL.DataFactory;
 
 /// <summary>
 /// Creates and sets up channels, roles and messages for the team.
 /// </summary>
+//TODO: JR - refactor
 public class DiscordTeam
 {
-    private static Dictionary<int, DiscordTeam> instances = new();
-
     private readonly DiscordClient discordClient;
     private readonly IDataWorker dataWorker = CreateDataWorker();
-    private readonly string changeTileButtonId = "{0}_change_tile_button";
-    private readonly string submitEvidenceButtonId = "{0}_submit_evidence_button";
-    private readonly string submitDropButtonId = "{0}_submit_drop_button";
-    private readonly string viewEvidenceButtonId = "{0}_view_evidence_button";
-    private readonly string clearEvidenceButtonId = "{0}_clear_evidence_button";
-    private readonly string completeNextTileEvidenceButtonId = "{0}_complete_next_tile_button";
+
+    #region channelNames
+
+    private const string categoryChannelName = "{0}";
+    private const string boardChannelName = "{0}-board";
+    private const string generalChannelName = "{0}-general";
+    private const string voiceChannelName = "{0}-voice";
+
+    #endregion
+
+    #region buttonIds
+
+    private const string changeTileButtonId = "{0}_change_tile_button";
+    private const string submitEvidenceButtonId = "{0}_submit_evidence_button";
+    private const string submitDropButtonId = "{0}_submit_drop_button";
+    private const string viewEvidenceButtonId = "{0}_view_evidence_button";
+    private const string clearEvidenceButtonId = "{0}_clear_evidence_button";
+    private const string completeNextTileEvidenceButtonId = "{0}_complete_next_tile_button";
+
+    #endregion
+
+    private static Dictionary<int, DiscordTeam> instances = new();
 
     private Team team = null!;
     private DiscordMessage boardMessage;
+
+    public string Name { get; private set; } = null!;
+    public DiscordRole Role { get; private set; } = null!;
+
+    public DiscordChannel BoardChannel { get; private set; } = null!;
+
+    public delegate DiscordTeam Factory(string name);
 
     public DiscordTeam(DiscordClient discordClient, string name)
     {
         Name = name;
         this.discordClient = discordClient;
-        SetButtonIds();
     }
-
-    public delegate DiscordTeam Factory(string name);
-
-    public string Name { get; private set; } = null!;
-
-    public DiscordChannel BoardChannel { get; private set; } = null!;
 
     public static async Task UpdateBoard(Team team, Image boardImage) =>
         await instances[team.RowId].UpdateBoardMessage(boardImage);
@@ -54,10 +69,12 @@ public class DiscordTeam
     {
         if (instances.ContainsKey(team.RowId))
         {
-            // TODO: JR - dispose of the instance
             instances.Remove(team.RowId);
         }
     }
+
+    public static DiscordTeam GetInstance(Team team) =>
+        instances[team.RowId];
 
     /// <summary>
     /// Creates and initializes the team's channel.
@@ -65,10 +82,14 @@ public class DiscordTeam
     /// </summary>
     public async Task InitialiseAsync()
     {
-        await CreateChannels();
-        await InitialiseBoardChannel();
-        CreateTeamEntry();
-        await UpdateBoardMessage(BoardImage.CreateBoard(team));
+        List<ulong> channelAndMessageIds = new(await CreateChannels())
+        {
+            await InitialiseBoardChannel(),
+            await CreateRole()
+        };
+
+        CreateTeamEntry(channelAndMessageIds);
+        await UpdateBoardMessage(BoardImage.Create(team));
 
         CommonInitialisation();
     }
@@ -82,6 +103,7 @@ public class DiscordTeam
         team = existingTeam;
         BoardChannel = await discordClient.GetChannelAsync(team.BoardChannelId);
         boardMessage = await BoardChannel.GetMessageAsync(team.BoardMessageId);
+        Role = Guild.GetRole(team.RoleId);
 
         CommonInitialisation();
     }
@@ -92,61 +114,65 @@ public class DiscordTeam
         RegisterBoardChannelComponentInteractions();
     }
 
-    private void SetButtonIds()
+    private async Task<List<ulong>> CreateChannels()
     {
-        changeTileButtonId.FormatConst(Name);
-        submitEvidenceButtonId.FormatConst(Name);
-        submitDropButtonId.FormatConst(Name);
-        viewEvidenceButtonId.FormatConst(Name);
-        clearEvidenceButtonId.FormatConst(Name);
-        completeNextTileEvidenceButtonId.FormatConst(Name);
+        List<ulong> ids = new(4);
+
+        DiscordChannel category = await Guild.CreateChannelAsync(GetId(categoryChannelName), ChannelType.Category);
+        BoardChannel = await Guild.CreateChannelAsync(GetId(boardChannelName), ChannelType.Text, category);
+
+        ids.Add(category.Id);
+        ids.Add(BoardChannel.Id);
+        ids.Add((await Guild.CreateChannelAsync(GetId(generalChannelName), ChannelType.Text, category)).Id);
+        ids.Add((await Guild.CreateChannelAsync(GetId(voiceChannelName), ChannelType.Voice, category)).Id);
+        return ids;
     }
 
-    private void CreateTeamEntry()
+    private async Task<ulong> CreateRole()
     {
-        team = TeamRecord.CreateTeam(dataWorker, Name, BoardChannel.Id, boardMessage.Id);
+        Role = await Guild.CreateRoleAsync(Name);
+        return Role.Id;
+    }
+
+    private void CreateTeamEntry(List<ulong> ids)
+    {
+        team = TeamRecord.CreateTeam(dataWorker, Name, ids[0], ids[1], ids[2], ids[3], ids[4], ids[5]);
         dataWorker.SaveChanges();
     }
 
-    private async Task CreateChannels()
-    {
-        DiscordChannel? category = await Guild.CreateChannelAsync($"{Name}", ChannelType.Category);
-        BoardChannel = await Guild.CreateChannelAsync($"{Name}-board", ChannelType.Text, category);
-        await Guild.CreateChannelAsync($"{Name}-general", ChannelType.Text, category);
-        await Guild.CreateChannelAsync($"{Name}-submitted-evidence", ChannelType.Text, category);
-        await Guild.CreateChannelAsync($"{Name}-voice", ChannelType.Voice, category);
-    }
+    private string GetId(string stringToFormat) =>
+        stringToFormat.FormatConst(Name);
 
     private async Task UpdateBoardMessage(Image boardImage)
     {
         var changeTileButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            changeTileButtonId,
+            GetId(changeTileButtonId),
             "Change tile");
 
         var submitEvidenceButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            submitEvidenceButtonId,
+            GetId(submitEvidenceButtonId),
             "Submit evidence");
 
         var submitDropButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            submitDropButtonId,
+            GetId(submitDropButtonId),
             "Submit drop");
 
         var viewEvidenceButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            viewEvidenceButtonId,
+            GetId(viewEvidenceButtonId),
             "View evidence");
 #if DEBUG
         var clearEvidenceButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            clearEvidenceButtonId,
+            GetId(clearEvidenceButtonId),
             "Clear evidence");
 
         var completeNextTileButton = new DiscordButtonComponent(
             ButtonStyle.Primary,
-            completeNextTileEvidenceButtonId,
+            GetId(completeNextTileEvidenceButtonId),
             "Complete next tile");
 #endif
 
@@ -157,7 +183,7 @@ public class DiscordTeam
         using (var fs = new FileStream(imageName, FileMode.Open, FileAccess.Read))
         {
             imageMessage = await BoardChannel.SendMessageAsync(new DiscordMessageBuilder()
-                .WithFile("Team board.png", fs));
+                .AddFile("Team board.png", fs));
         }
 
         var boardImageEmbed = new DiscordEmbedBuilder()
@@ -168,34 +194,52 @@ public class DiscordTeam
         .Build();
 
         var builder = new DiscordMessageBuilder()
-            .WithEmbed(boardImageEmbed)
-            .AddComponents(changeTileButton, submitEvidenceButton, submitDropButton, viewEvidenceButton)
-            .AddComponents(clearEvidenceButton, completeNextTileButton);
+            .WithEmbed(boardImageEmbed);
+
+        if (EnableBoardCustomisation)
+        {
+            builder.AddComponents(changeTileButton, submitEvidenceButton, submitDropButton, viewEvidenceButton);
+        }
+        else
+        {
+            builder.AddComponents(submitDropButton, viewEvidenceButton);
+        }
+
+#if DEBUG
+        builder.AddComponents(clearEvidenceButton, completeNextTileButton);
+#endif
+
         // TODO: JR - get this to work with a file upload instead of an embed since it will look nicer.
         //.WithFile("Team board.png", fs, true);
         await boardMessage.ModifyAsync(builder);
         await imageMessage.DeleteAsync();
     }
 
-    private async Task InitialiseBoardChannel()
+    private async Task<ulong> InitialiseBoardChannel()
     {
         var builder = new DiscordMessageBuilder()
             .WithContent("Loading...");
         boardMessage = await BoardChannel.SendMessageAsync(builder);
+        return boardMessage.Id;
     }
 
     private void RegisterBoardChannelComponentInteractions()
     {
-        ComponentInteractionHandler.InitialisationInfo info = new ()
+        ComponentInteractionHandler.InitialisationInfo info = new()
         {
             Team = this,
         };
-        
-        ComponentInteractionHandler.Register<ChangeTileButtonHandler>(changeTileButtonId, info);
-        ComponentInteractionHandler.Register<SubmitEvidenceButtonHandler>(submitEvidenceButtonId, info); 
-        ComponentInteractionHandler.Register<SubmitDropButtonHandler>(submitDropButtonId, info);
-        ComponentInteractionHandler.Register<ViewEvidenceButtonHandler>(viewEvidenceButtonId, info);
-        ComponentInteractionHandler.Register<ClearTeamsEvidenceButtonHandler>(clearEvidenceButtonId, info);
-        ComponentInteractionHandler.Register<CompleteNextTileButtonHandler>(completeNextTileEvidenceButtonId, info);
+
+        ComponentInteractionHandler.Register<ChangeTileButtonHandler>(GetId(changeTileButtonId), info);
+        ComponentInteractionHandler.Register<SubmitEvidenceButtonHandler>(GetId(submitEvidenceButtonId), info);
+        ComponentInteractionHandler.Register<SubmitDropButtonHandler>(GetId(submitDropButtonId), info);
+        ComponentInteractionHandler.Register<ViewEvidenceButtonHandler>(GetId(viewEvidenceButtonId), info);
+
+#if DEBUG
+
+        ComponentInteractionHandler.Register<ClearTeamsEvidenceButtonHandler>(GetId(clearEvidenceButtonId), info);
+        ComponentInteractionHandler.Register<CompleteNextTileButtonHandler>(GetId(completeNextTileEvidenceButtonId), info);
+
+#endif
     }
 }
