@@ -4,283 +4,81 @@
 
 namespace RSBingoBot.Imaging;
 
-using RSBingo_Framework.Interfaces;
 using RSBingo_Framework.Models;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
 using SixLabors.ImageSharp;
-using static RSBingo_Framework.DAL.DataFactory;
-using SixLabors.ImageSharp.Processing;
-using RSBingo_Framework;
 using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.ImageSharp.Drawing.Processing;
-using SixLabors.Fonts;
+using SixLabors.ImageSharp.Processing;
 using static RSBingo_Common.General;
-using static BoardPreferences;
-using static RSBingo_Framework.Records.BingoTaskRecord;
-using RSBingo_Framework.Exceptions;
+using static RSBingoBot.Imaging.BoardPreferences;
 
-public class BoardImage
+internal static class BoardImage
 {
-    private const string boardBaseFileName = "Board base.png";
-    private const string noTaskName = "No task";
-
-    private static readonly IDataWorker DataWorker;
-    private static readonly Dictionary<int, TaskInfo> tasksInfo = new();
-    private static readonly List<Image> baseTiles = new();
-    private static readonly Dictionary<int, Board> boards = new();
-    private static readonly Font font;
-    private static readonly TextOptions textOptions;
-    private static readonly Image<Rgba32> emptyTaskImage = new(TilePixelWidth, TileBorderPixelHeight);
-
-    private static Image<Rgba32> boardBase = null!;
+    private static Image baseBoard;
 
     static BoardImage()
     {
-        FontCollection collection = new();
-        FontFamily family = collection.Add(FontPath);
-        font = family.CreateFont(FontSize, Style);
-
-        DataWorker = CreateDataWorker();
-        DisectBoardBase();
-
-        textOptions = new(font)
-        {
-            Origin = new PointF(baseTiles[0].Width / 2, TextTopOffsetPixels),
-            WrappingLength = TilePixelWidth - TextXPaddingPixels,
-            HorizontalAlignment = HorizontalAlignment.Center,
-            VerticalAlignment = VerticalAlignment.Top,
-            TextAlignment = TextAlignment.Center
-        };
+        baseBoard = Image.Load(GetBaseBoardImagePath());
     }
 
-    /// <summary>
-    /// Creates all boards for existing teams and updates them in their channels. Currently just used to test how the board looks.
-    /// </summary>
-    /// <returns></returns>
-    public static async Task CreateAndUpdateAllTeamBoards()
+    public static Image Create(Team team)
     {
-        foreach (Team team in DataWorker.Teams.GetAll())
-        {
-            await DiscordTeam.UpdateBoard(team, CreateBoard(team));
-        }
-    }
-
-    public static Image CreateBoard(Team team)
-    {
-        boards.Add(team.RowId, new(team));
+        Image teamBoard = GetBoard(team);
 
         foreach (Tile tile in team.Tiles)
         {
-            UpdateTile(team, tile.BoardIndex, tile.Task);
+            UpdateTile(teamBoard, tile);
         }
 
-        return boards[team.RowId].Image;
+        return teamBoard;
     }
 
-    public static Image GetTeamBoard(Team team) =>
-        boards.ContainsKey(team.RowId) ?
-            boards[team.RowId].Image :
-            CreateBoard(team);
-
-    public static Image<Rgba32> UpdateTile(Team team, int tileIndex, BingoTask? task)
+    public static Image UpdateTile(Image board, Tile tile)
     {
-        if (!boards.ContainsKey(team.RowId))
-        {
-            CreateBoard(team);
-        }
+        Rectangle tileRect = GetTileRect(tile.BoardIndex);
+        Point taskImagePosition = new(tileRect.Width / 2 + TaskXOffsetPixels, tileRect.Height / 2 + TaskYOffsetPixels);
 
-        UpdateTaskInfo(team, tileIndex, task);
-        UpdateBoardImage(boards[team.RowId], team, tileIndex);
+        Image taskImage = GetTaskImage(tile.Task.Name, tileRect);
+        Image tileImage = baseBoard.Clone(b => b.Crop(tileRect));
+        tileImage.Mutate(t => t.DrawImage(taskImage, taskImagePosition, 1));
 
-        return boards[team.RowId].Image;
+        board.Mutate(b => b.DrawImage(tileImage, new Point(tileRect.X, tileRect.Y), 1));
+
+        return board;
     }
 
-    private static TaskInfo UpdateTaskInfo(Team team, int tileIndex, BingoTask? task)
+    public static Image UpdateTile(Tile tile)
     {
-        TaskInfo taskInfo;
-
-        if (task == null)
-        {
-            taskInfo = new();
-        }
-        else
-        {
-            GetTaskInfo(task);
-            taskInfo = tasksInfo[task.RowId];
-        }
-
-        boards[team.RowId].Tiles[tileIndex].taskInfo = taskInfo;
-        return taskInfo;
+        string teamBoardPath = GetBoardImagePath(tile.Team.Name);
+        Image teamBoard = Image<Rgba32>.Load(teamBoardPath);
+        return UpdateTile(teamBoard, tile);
     }
 
-    private static Image<Rgba32> UpdateBoardImage(Board board, Team team, int tileIndex)
+    public static void ClearTile(Image board, Tile tile)
     {
-        TileInfo tileInfo = board.Tiles[tileIndex];
-        TaskInfo taskInfo = tileInfo.taskInfo;
-
-        Image tileImage = baseTiles[tileIndex]
-            .Clone(i => i.DrawImage(taskInfo.Image, new Point(taskInfo.XPadding, taskInfo.YPadding), 1));
-
-        DrawTaskNameOnImage(tileImage, taskInfo.Name);
-        board.Image.Mutate(i => i.DrawImage(tileImage, new Point(tileInfo.x, tileInfo.y), 1));
-        return board.Image;
+        Rectangle tileRect = GetTileRect(tile.BoardIndex);
+        Image tileImage = baseBoard.Clone(b => b.Crop(tileRect));
+        board.Mutate(b => b.DrawImage(tileImage, new Point(tileRect.X, tileRect.Y), 1));
     }
 
-    private static void DrawTaskNameOnImage(Image image, string name) =>
-        image.Mutate(i => i.DrawText(textOptions, name, Color.Black));
-
-    private static void DisectBoardBase()
+    public static Image GetBoard(Team team)
     {
-        // TODO: JR - have the dissection info stored in either a config file or the DB, and have it be changed via a command.
-
-        try
-        {
-            boardBase = Image.Load<Rgba32>(TaskTemplatePopulator.GetFilePath(boardBaseFileName));
-        }
-        catch
-        {
-            throw new BoardImageException($"Could not find the '{boardBaseFileName}' file. " +
-                "Please ensure one exists in order to create the team boards.");
-        }
-
-        CheckBoardBaseDimensions(boardBase);
-
-        for (int i = 0; i < TilesPerColumn; i++)
-        {
-            for (int j = 0; j < TilesPerRow; j++)
-            {
-                int xPos = GetTileXPosition(j);
-                int yPos = GetTileYPosition(i);
-                var rectangle = new Rectangle(xPos, yPos, TilePixelWidth, TilePixelHeight);
-                baseTiles.Add(boardBase.Clone(i => i.Crop(rectangle)));
-            }
-        }
+        string teamBoardPath = GetBoardImagePath(team.Name);
+        return File.Exists(teamBoardPath) ? Image<Rgba32>.Load(teamBoardPath) : baseBoard.Clone(b => { });
     }
 
-    private static void CheckBoardBaseDimensions(Image boardBase)
+    private static Image GetTaskImage(string taskName, Rectangle tileRect)
     {
-        int expectedBoardBaseWidth = BoardBorderPixelWidth * 2 +
-            TilePixelWidth * TilesPerRow +
-            TileBorderPixelWidth * (TilesPerRow - 1);
-
-        if (boardBase.Width != expectedBoardBaseWidth)
-        {
-            throw new BoardImageException($"Incorrect board base width. " +
-                $"Expected {expectedBoardBaseWidth} but was {boardBase.Width}.");
-        }
-
-        int expectedBoardBaseHeight = BoardBorderPixelHeight * 2 +
-            TilePixelHeight * TilesPerColumn +
-            TileBorderPixelHeight * (TilesPerColumn - 1);
-
-        if (boardBase.Height != expectedBoardBaseHeight)
-        {
-            throw new BoardImageException($"Incorrect board base height. " +
-                $"Expected {expectedBoardBaseHeight} but was {boardBase.Height}.");
-        }
+        // TODO: resize all images on startup
+        Image taskImage = Image<Rgba32>.Load(GetTaskImagePath(taskName));
+        Size size = new Size(tileRect.Width - TaskXPaddingPixels * 2, tileRect.Height - TaskYPaddingPixels * 2);
+        taskImage.Mutate(i => i.Resize(size));
+        return taskImage;
     }
 
-    private static TaskInfo GetTaskInfo(BingoTask? task)
+    private static Rectangle GetTileRect(int tileIndex)
     {
-        if (task == null) { return new(); }
-
-        // Check if it's already loaded.
-        if (tasksInfo.ContainsKey(task.RowId)) { return tasksInfo[task.RowId]; }
-
-        Image image;
-        string imagePath = GetTaskImagePath(task.Name);
-        string taskName = task != null ? task.Name : noTaskName;
-        int taskId = task != null ? task.RowId : -1;
-
-        try
-        {
-            image = Image.Load<Rgba32>(imagePath);
-        }
-        catch
-        {
-            LoggingLog($"Could not load the {task!.Name} task image, it may have been corrupted; moved or deleted.");
-            image = emptyTaskImage;
-            taskName = "Unable to find image";
-        }
-
-        var resizeOptions = new ResizeOptions()
-        {
-            Mode = ResizeMode.Max,
-            Size = new Size(
-                TilePixelWidth - TaskXPaddingPixels * 2 - (int)MathF.Abs(TaskXOffsetPixels),
-                TilePixelHeight - TaskYPaddingPixels * 2 - (int)MathF.Abs(TaskYOffsetPixels))
-        };
-
-        image.Mutate(i => i.Resize(resizeOptions));
-
-        int x = (TilePixelWidth - image.Width + TaskXOffsetPixels) / 2;
-        int y = (TilePixelHeight - image.Height + TaskYOffsetPixels) / 2;
-
-        TaskInfo taskInfo = new(image, taskName, x, y);
-        tasksInfo.Add(taskId, taskInfo);
-        return taskInfo;
-    }
-
-    private static int GetTileXPosition(int tileRowIndex) =>
-        BoardBorderPixelWidth + (TilePixelWidth + TileBorderPixelWidth) * tileRowIndex;
-
-    private static int GetTileYPosition(int tileColumnIndex) =>
-        BoardBorderPixelHeight + (TilePixelHeight + TileBorderPixelHeight) * tileColumnIndex;
-
-    private class TaskInfo
-    {
-        public Image Image { get; }
-        public string Name { get; }
-        public int XPadding { get; }
-        public int YPadding { get; }
-
-        public TaskInfo()
-        {
-            Image = emptyTaskImage;
-            Name = noTaskName;
-            XPadding = 0;
-            YPadding = 0;
-        }
-
-        public TaskInfo(Image image, string name, int xPadding, int yPadding)
-        {
-            Image = image;
-            Name = name;
-            XPadding = xPadding;
-            YPadding = yPadding;
-        }
-    }
-
-    private class TileInfo
-    {
-        public TaskInfo taskInfo { get; set; }
-        public int baseTileIndex { get; }
-        public int x { get; }
-        public int y { get; }
-
-        public TileInfo(TaskInfo taskInfo, int baseTileIndex, int x, int y)
-        {
-            this.taskInfo = taskInfo;
-            this.baseTileIndex = baseTileIndex;
-            this.x = x;
-            this.y = y;
-        }
-    }
-
-    private class Board
-    {
-        public Image<Rgba32> Image = boardBase.Clone();
-        public TileInfo[] Tiles = new TileInfo[MaxTilesOnABoard];
-
-        public Board(Team team)
-        {
-            for (int i = 0; i < MaxTilesOnABoard; i++)
-            {
-                Tiles[i] = new(new TaskInfo(), i, GetTileXPosition(i % TilesPerRow), GetTileYPosition(i / TilesPerColumn));
-                UpdateBoardImage(this, team, i);
-            }
-        }
+        int x = BoardBorderPixelWidth + (TilePixelWidth + TileBorderPixelWidth) * tileIndex;
+        int y = BoardBorderPixelHeight + (TilePixelHeight + TileBorderPixelHeight) * tileIndex;
+        return new(x, y, TilePixelWidth, TilePixelHeight);
     }
 }
